+++ apps/isomorphic-i18n/src/app/shared/file/dashboard/charts/catch-time-series.tsx
@@ -218,7 +218,7 @@ export default function CatchTimeSeries({
               return (
                 <Line
                   key={district}
-                  type="monotone"
+                  type="linear"
                   dataKey={district}
                   stroke={color}
                   strokeWidth={3}